command/{app, cp, options}, storage/: updates based on pr review
